Moved `host/syscall_condition.c` to `host/syscall/syscall_condition.c`
