+++ src/main/host/syscall/futex.c
@@ -16,7 +16,7 @@
 #include "main/host/futex.h"
 #include "main/host/futex_table.h"
 #include "main/host/syscall/protected.h"
-#include "main/host/syscall_condition.h"
+#include "main/host/syscall/syscall_condition.h"
 #include "main/utility/utility.h"
 
 ///////////////////////////////////////////////////////////
