+++ src/main/host/syscall/uio.c
@@ -17,7 +17,7 @@
 #include "main/host/process.h"
 #include "main/host/syscall/protected.h"
 #include "main/host/syscall/unistd.h"
-#include "main/host/syscall_condition.h"
+#include "main/host/syscall/syscall_condition.h"
 
 ///////////////////////////////////////////////////////////
 // Helpers
