+++ src/main/host/syscall/unistd.c
@@ -16,7 +16,7 @@
 #include "main/host/descriptor/regular_file.h"
 #include "main/host/process.h"
 #include "main/host/syscall/protected.h"
-#include "main/host/syscall_condition.h"
+#include "main/host/syscall/syscall_condition.h"
 #include "main/host/syscall_handler.h"
 
 ///////////////////////////////////////////////////////////
